requirements.txt
anyio
paho-mqtt
python-slugify